requirements.txt
# 核心依赖
requests>=2.28.0
pymysql>=1.0.0
sqlalchemy>=1.4.0

# Telegram Bot
python-telegram-bot>=20.0

# 异步处理 (可选)
asyncio-mqtt>=0.11.0

# 日志和监控 (Python内置)

# 加密
cryptography>=3.4.0

# 时间处理
python-dateutil>=2.8.0

# 配置管理
python-dotenv>=0.19.0
